ek_haseena/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
Pillow>=10.0.0
pytesseract>=0.3.10
cryptography>=42.0.0
